Improve loading of boundary and selection layers.
There's a JS race condition in effect if you load the boundary and selection (the layer that shows what's currently selected) layers. Both were set to `position: "front"` so whichever loaded last won out, making the selection outline drop behind the boundary layer sometimes.. This change makes a guess at a different `position` keyword for the layer that we don't want to end up on top, though the documentation on it seems to suggest that "front" is the only value. :)