Load geographies from API.
For simplicity, the geographies were stored in the JS file. However, we're now actively adding to that list, so need to load them at load.

Much of the loader plumbing had to be reorganized to wait until the geography API call was completed, because the map and Vuw instances relied on knowing what the current geography was.

Also, the geography selector above the map us now handled in Vue, after the geographies are fetched.